version2，增加了idgen，策略解析，多协程处理，实现从订单创建到结果查询

+++ model/device_status.go
@@ -66,7 +66,7 @@ func (d *DeviceStatusService)GetAllInfo() []*DeviceStatus{
 }
 
 //查询一个device id下的所有记录
-func (d *DeviceStatusService)GetStatusByDeviceID(deviceID int) []*DeviceStatus{
+func (d *DeviceStatusService)GetStatusByDeviceID(deviceID uint64) []*DeviceStatus{
 	var DevStatusList []*DeviceStatus
 	err := DBIot.Where("deviceid=?",deviceID).Find(&DevStatusList).Error
 	if err != nil{
